change: add flexbox with new containers to modal
Separate the contents of the alert box into smaller containers, this way
they can be manipulated with flexbox tools to reach the desired layout.

Note:
- If I have a flex container with several flex items but I want the
container to exclude one or more from its properties, I can nest these
in a <div>. I did this with the continue and cancel buttons so that
they preserved their original properties.